+++ CHANGELOG.md
@@ -1,2 +1,6 @@
-v1.0.0:
-- initial
+v1.0.1:
+### Changes
+- Added option `mod_enabled` to config:
+    - Turn borrowing off entirely
+- Added option `empty_hand_only` to config:
+    - Borrow only when your main hand are empty. Keeps the mod from item interactions.